very_main

+++ C/C02/ex01/ft_strncpy.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strncpy.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: amura <marvin@42.fr>                       +#+  +:+       +#+        */
+/*   By: amura <amura@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/11 10:34:30 by amura             #+#    #+#             */
-/*   Updated: 2023/08/11 10:41:14 by amura            ###   ########.fr       */
+/*   Updated: 2023/08/14 10:23:37 by amura            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 char	*ft_strncpy(char *dest, char *src, unsigned int n)
 {
-	size_t		i;
+	unsigned int	i;
 
 	i = 0;
 	while (i < n && src[i] != '\0')
